+++ core/src/main/kotlin/io/heapy/kotbot/bot/_kotbot.kt
@@ -1,5 +1,7 @@
 package io.heapy.kotbot.bot
 
+import io.ktor.client.HttpClient
+import io.ktor.client.engine.cio.CIO
 import io.ktor.client.request.header
 import io.ktor.client.request.post
 import io.ktor.client.request.setBody
@@ -8,15 +10,69 @@ import io.ktor.client.statement.bodyAsText
 import io.ktor.http.ContentType
 import io.ktor.http.HttpHeaders
 import io.ktor.http.isSuccess
+import kotlinx.coroutines.flow.Flow
+import kotlinx.coroutines.flow.flow
 import kotlinx.serialization.Serializable
+import kotlinx.serialization.json.Json
+import org.slf4j.Logger
+import org.slf4j.LoggerFactory
 import java.lang.RuntimeException
 
-public interface ApiMethod<Response> {
+/**
+ * Workaround until [multi-receivers](https://youtrack.jetbrains.com/issue/KT-10468) will be implemented in Kotlin.
+ */
+public data class Kotbot(
+    public val token: String,
+    public val baseUrl: String = "https://api.telegram.org/bot",
+    public val httpClient: HttpClient = HttpClient(CIO) {
+        engine {
+            requestTimeout = 60_000
+        }
+    },
+    public val json: Json = Json {
+        ignoreUnknownKeys = true
+    }
+)
+
+public suspend inline fun <Response> Kotbot.execute(
+    method: Method<Response>
+): Response =
+    with(method) {
+        execute()
+    }
+
+public suspend fun Kotbot.receiveUpdates(
+    timeout: Int = 50,
+    limit: Int = 100,
+    allowedUpdates: List<String> = listOf(),
+): Flow<Update> {
+    var offset: Int? = null
+
+    return flow {
+        while (true) {
+            val updates = execute(GetUpdates(
+                offset = offset?.let { it + 1 },
+                limit = limit,
+                timeout = timeout,
+                allowed_updates = allowedUpdates,
+            ))
+
+            offset = updates.maxByOrNull { it.update_id }?.update_id ?: offset
+
+            updates.forEach { update ->
+                LOGGER.debug("Received update: {}", update)
+                emit(update)
+            }
+        }
+    }
+}
+
+public interface Method<Response> {
     public suspend fun Kotbot.execute(): Response
 }
 
 @Serializable
-public data class ApiResponse<Result>(
+public data class Response<Result>(
     public val ok: Boolean,
     public val result: Result? = null,
     public val error_code: Int? = null,
@@ -30,7 +86,7 @@ public data class ResponseParameters(
     public val retry_after: Int?,
 )
 
-internal fun <T> ApiResponse<T>.unwrap(): T {
+internal fun <T> Response<T>.unwrap(): T {
     if (ok) {
         return result ?: throw KotbotException("Response is ok but result is null")
     } else {
@@ -73,3 +129,8 @@ internal suspend inline fun <Response> Kotbot.requestForJson(
         throw TelegramApiError("${response.status} ${response.bodyAsText()}")
     }
 }
+
+private inline fun <reified T : Any> logger(): Logger =
+    LoggerFactory.getLogger(T::class.java)
+
+private val LOGGER = logger<Kotbot>()
