+++ core/src/main/kotlin/io/heapy/kotbot/bot/_model.kt
