+++ core/src/main/kotlin/io/heapy/kotbot/bot/deleteWebhook.kt
@@ -16,10 +16,10 @@ public data class DeleteWebhook(
      * Pass True to drop all pending updates
      */
     val drop_pending_updates: Boolean? = null,
-) : ApiMethod<Boolean> {
+) : Method<Boolean> {
     @Transient
-    private val deserializer: KSerializer<ApiResponse<Boolean>> =
-        ApiResponse.serializer(Boolean.serializer())
+    private val deserializer: KSerializer<Response<Boolean>> =
+        Response.serializer(Boolean.serializer())
 
     override suspend fun Kotbot.execute(): Boolean {
         return requestForJson(
